Add employer metadata list to preloaded state
